Use formspree.io AJAX json API for contact form
Without this patch the contact form requires server side processing of a
PHP file.  This is a problem because we're using a static site hosting
service on github pages.  This patch addresses the problem by sending
the contact form data to formspree.io using AJAX.

Formspree is interesting because it can also be deployed to Heroku
according to the documentation at
https://github.com/asm-products/formspree