Configure https://openinfrastructure.co on core1
This gets everything ready for DNS redirection of the apex and www
domains.

Note, redirect cloudflare in proxy mode using the
ap@openinfrastructure.co login to get to the domains.

Encryption is set to full mode and the cert issued is a cluster managed
cert instead of a letsencrypt cert.  This allows cloudflare to manage
the public cert.

Closes: #3