Add Action to deploy GH Pages
PROBLEM:

GitHub has a paved path for deploying Jekyll sites, but with the
switch to using the Hugo/Docsy site we will need automation to deploy
the site.

SOLUTION:

Use the GH Action provided by Hugo to deploy to GH Pages on commits to
the `main` branch.

See: https://gohugo.io/hosting-and-deployment/hosting-on-github/

OUTCOME:

We are deploying to GH Pages every time a new commit lands on the `main`
branch.