Upgrade Docsy and fix formatting
PROBLEM:

The site has been using an older version of Hugo and Docsy. When we
tried to add the new Holos page and rendered the site using the newest
versions of Hugo/Docsy, lots of formatting and page elements broke.

SOLUTION:

* Update site configuration to specify the new versions of Docsy/Hugo.
* Follow the [site conversion/migration guide](https://www.docsy.dev/docs/updating/convert-site-to-module/) to use Docsy as a Go module.
* Convert all image symlinks to native files, as Hugo no longer resolves the symlinks.
* Update formatting in the `about` page to use row/column shortcodes.
* Specify the `row` type on the main homepage so the feature blocks align as columns.

OUTCOME:

Executing `hugo serve` locally produces a website with formatting that
matches our existing site.