Fix horizontal columns in About
PROBLEM:

The columns in the "about" page weren't expanding the width of the page.
This is because of a clash between the `row` div in the `blocks/section`
shortcode and the `row` shortcode.

SOLUTION:

Allow the `blocks/section` shortcode to set the `row` div itself, and remove
the `row` shortcode. This gives us 100% page width. We also need to set
the custom `book` ID to allow text beside the images.

OUTCOME:

Alignment on the About page matches what we had before.