Refactor the site to use Docusaurus
PROBLEM:

The existing site is using out of date versions of Docsy and Hugo. We
need to either upgrade those components, or use the time to switch to
a different static site generator.

SOLUTION:

Use Docusaurus. Jeff has had a good experience using it with Holos, so
let's use the opportunity to switch the site to Docusaurus.

This commit refactors the existing site content to use Docusaurus with
the following details:

Removed the Logo blog as it really doesn't serve us well.
Used undraw.co SVG graphics on the landing page.
Removed "Google Partners" because evidently we aren't anymore.
Setup redirects to the existing site URLs so Google won't complain.
OUTCOME:

Our existing site content, but on a different static site generation
platform, and one that is easy for us to update.